palette, gamepad, primitive, collider
